# openapi.yaml
openapi: 3.1.0
info:
  title: FindTutor API
  version: 0.1.0
  description: >
    Contract for the FindTutor language-tutoring marketplace (Uzbekistan).
    Hand-written and treated as the source of truth: the Go backend renders
    responses to match it, and the Next.js frontend generates its typed client
    from it (openapi-typescript + openapi-fetch).

    Money is always an integer amount in MINOR units (tiyin for UZS, cents for
    USD) plus a currency code — never a float.

    Language: every human-readable `error.message` is rendered in the language
    of the request's `Accept-Language` header (`en`, `ru` or `uz`; anything
    else falls back to English) and the response carries `Content-Language`.
    Email is written in the account's stored `locale` (see AuthUser), set from
    Accept-Language at registration and changed via PATCH /v1/auth/me.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/teachers:
    get:
      tags: [teachers]
      summary: Search teachers
      operationId: listTeachers
      parameters:
        - name: language
          in: query
          description: Filter to teachers who teach this language (ISO 639-1 code, e.g. `en`).
          schema: { type: string }
        - name: kind
          in: query
          schema: { $ref: "#/components/schemas/TeacherKind" }
        - name: max_price_minor
          in: query
          description: Only teachers whose hourly price is at or below this amount (minor units).
          schema: { type: integer, format: int64, minimum: 0 }
        - name: q
          in: query
          description: Case-insensitive substring match across name, headline, and focus tags.
          schema: { type: string }
        - name: sort
          in: query
          schema:
            $ref: "#/components/schemas/TeacherSort"
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        "200":
          description: A page of matching teachers plus catalog-wide facets.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeacherListResult" }
        "400":
          $ref: "#/components/responses/BadRequest"
    post:
      tags: [teachers]
      summary: Create the caller's teacher profile
      description: >
        Claims a teacher profile for the authenticated account. One profile per
        account — a second attempt is 409. The new profile's owner is the caller.

        The body carries only the editable fields; server-controlled aggregates
        (`rating`, `review_count`, `lessons_completed`, `student_count`,
        `response_time_hours`) start at 0, `accepting_students` starts true, and
        the `slug` is generated from `display_name` (de-duplicated with a numeric
        suffix). `languages` / `focus` / `experience` are stored as given.


        On create, `display_name`, `headline`, `kind`, `country_code`,
        `country_name`, `city`, `timezone`, and `price_per_hour_minor` are
        required; a missing or blank one is a 400.
      operationId: createTeacherProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TeacherProfileWritable" }
      responses:
        "201":
          description: The created profile (same shape as GET /v1/teachers/{slug}).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeacherProfile" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: The account already has a teacher profile.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /v1/teachers/{slug}:
    get:
      tags: [teachers]
      summary: Get one teacher profile
      operationId: getTeacherBySlug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The full teacher profile.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeacherProfile" }
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [teachers]
      summary: Edit the caller's own teacher profile
      description: >
        Partial update of the profile at `slug`. Requires a Bearer token whose
        account owns the profile (`teachers.user_id`); otherwise 401 (no/invalid
        token) or 403 (not the owner — an unclaimed profile is never editable).

        Every field is optional; an omitted field is left unchanged. A present
        `languages` / `focus` / `experience` array (even `[]`) fully replaces
        that collection. The `slug` is immutable and server-controlled
        aggregates cannot be set here.
      operationId: updateTeacherProfile
      security:
        - bearerAuth: []
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TeacherProfileWritable" }
      responses:
        "200":
          description: The updated profile.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TeacherProfile" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/bookings:
    post:
      tags: [bookings]
      summary: Book a lesson
      description: >
        Creates a booking in `pending_payment` and opens a `requires_payment`
        payment intent for its price. The server re-derives the bookable slot
        set and rejects a `start_at` that is not a real, currently free slot; it
        never trusts a client-supplied price.

        `duration_minutes` must be one of 30/60/90/120. When `is_trial` is true
        the duration is forced to 30 minutes and the price is the teacher's
        `trial_price_minor` (400 if the teacher offers no trial). You cannot
        book your own teacher profile.

        The student then calls `POST /v1/bookings/{id}/pay` to authorise
        payment, which confirms the booking.
      operationId: createBooking
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateBookingRequest" }
      responses:
        "201":
          description: The created booking.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: >
            The slot is not bookable (`slot_unavailable` — outside availability,
            misaligned, in the past, or already taken) or was taken by a
            concurrent request between the check and the insert (`slot_taken`).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    get:
      tags: [bookings]
      summary: List the caller's bookings
      description: >
        With no `role`, returns every booking the caller participates in (as
        student or as the owner of the teacher profile). `role=student` limits
        it to bookings the caller booked; `role=teacher` to bookings for the
        teacher profile the caller owns (empty if they own none). Newest lesson
        first. Optional `status` filter.
      operationId: listBookings
      security:
        - bearerAuth: []
      parameters:
        - name: role
          in: query
          required: false
          schema: { type: string, enum: [student, teacher] }
        - name: status
          in: query
          required: false
          schema: { $ref: "#/components/schemas/BookingStatus" }
      responses:
        "200":
          description: The matching bookings.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BookingList" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/bookings/{id}:
    get:
      tags: [bookings]
      summary: Get one booking
      operationId: getBooking
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: The booking.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /v1/admin/users:
    get:
      tags: [admin]
      summary: User directory
      description: "Permission: `users.view`. `q` is a case-insensitive substring on email or display_name."
      operationId: adminListUsers
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1, minimum: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        "200":
          description: A page of users.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/admin/users/{id}:
    get:
      tags: [admin]
      summary: User detail
      description: "Permission: `users.view`. Bookings capped at the 50 newest."
      operationId: adminGetUser
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The user with roles, teacher profile, recent bookings, and a payments summary.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminUserDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /v1/admin/teachers:
    get:
      tags: [admin]
      summary: Teacher moderation queue
      description: "Permission: `teachers.view`. `status` filter optional; `q` matches display_name / slug / owner email."
      operationId: adminListTeachers
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { type: string, enum: [pending, approved, rejected, suspended] } }
        - { name: q, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        "200":
          description: A page of teacher profiles across all statuses.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminTeacherList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/admin/teachers/{slug}:
    get:
      tags: [admin]
      summary: Teacher detail (moderation)
      description: "Permission: `teachers.view`. The full public profile shape plus `status`, `verified`, `moderation_note`, and `owner`."
      operationId: adminGetTeacher
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: The teacher profile with its moderation slice.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminTeacherDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/admin/teachers/{slug}/approve:
    post:
      tags: [admin]
      summary: Approve a teacher
      description: >
        Permission: `teachers.moderate`. Sets `status = approved` and clears the
        moderation note. Allowed from pending, rejected, or suspended (the last is
        a reinstate). 409 `invalid_transition` from approved.
      operationId: adminApproveTeacher
      security: [{ bearerAuth: [] }]
      parameters: [{ name: slug, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: The updated teacher detail., content: { application/json: { schema: { $ref: "#/components/schemas/AdminTeacherDetail" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "`invalid_transition`."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /v1/admin/teachers/{slug}/reject:
    post:
      tags: [admin]
      summary: Reject a pending teacher
      description: "Permission: `teachers.moderate`. Body `{note}` required (non-empty). Allowed only from pending."
      operationId: adminRejectTeacher
      security: [{ bearerAuth: [] }]
      parameters: [{ name: slug, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [note], properties: { note: { type: string, minLength: 1 } } }
      responses:
        "200": { description: The updated teacher detail., content: { application/json: { schema: { $ref: "#/components/schemas/AdminTeacherDetail" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "`invalid_transition` — not in `pending`."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /v1/admin/teachers/{slug}/suspend:
    post:
      tags: [admin]
      summary: Suspend an approved teacher
      description: >
        Permission: `teachers.moderate`. Body `{note}` required. Allowed only from
        approved. Pulls the profile from public search immediately; existing
        bookings are untouched.
      operationId: adminSuspendTeacher
      security: [{ bearerAuth: [] }]
      parameters: [{ name: slug, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [note], properties: { note: { type: string, minLength: 1 } } }
      responses:
        "200": { description: The updated teacher detail., content: { application/json: { schema: { $ref: "#/components/schemas/AdminTeacherDetail" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "`invalid_transition` — not in `approved`."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /v1/admin/teachers/{slug}/verify:
    post:
      tags: [admin]
      summary: Set the verified badge
      description: "Permission: `teachers.verify`. Body `{verified: bool}`. Independent of moderation status."
      operationId: adminVerifyTeacher
      security: [{ bearerAuth: [] }]
      parameters: [{ name: slug, in: path, required: true, schema: { type: string } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [verified], properties: { verified: { type: boolean } } }
      responses:
        "200": { description: The updated teacher detail., content: { application/json: { schema: { $ref: "#/components/schemas/AdminTeacherDetail" } } } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/admin/bookings:
    get:
      tags: [admin]
      summary: Every booking on the platform
      description: >
        Permission: `bookings.view`. Unlike `GET /v1/bookings` (the caller's own
        lessons) this lists every booking, newest lesson first. `status` is an
        optional exact filter; `q` matches the teacher's display name or slug and
        the student's email or display name, case-insensitively.
      operationId: adminListBookings
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: status, in: query, schema: { $ref: "#/components/schemas/BookingStatus" } }
        - { name: q, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: page_size, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        "200":
          description: A page of bookings.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminBookingList" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /v1/admin/bookings/{id}:
    get:
      tags: [admin]
      summary: Booking detail (operator view)
      description: >
        Permission: `bookings.view`. The list-row fields plus the payment detail,
        the effective meeting link, the no-show flag, the cancellation
        who/why/when, and the booking's full dispute thread. The meeting link is
        shown regardless of booking status — unlike the participant-facing
        `Booking`, which hides it before confirmation.
      operationId: adminGetBooking
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: The booking.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminBookingDetail" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: "`booking_not_found`."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  /v1/admin/bookings/{id}/force-cancel:
    post:
      tags: [admin]
      summary: Force-cancel a booking
      description: >
        Permission: `bookings.force_cancel`. The operator override of
        `POST /v1/bookings/{id}/cancel`, which is participant-only: it cancels a
        `pending_payment` or `confirmed` booking on anyone's behalf and records
        `cancelled_by = admin` plus the required `reason`. 409 `invalid_state`
        for a booking that is already completed or cancelled.

        It runs the same cancellation path a participant cancel does — pending
        lesson reminders are dropped and the participants are emailed — and with
        `refund: true` the booking's payment is released (an authorized hold) or
        refunded (a capture). With `refund: false` no payment call is made.
      operationId: adminForceCancelBooking
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ForceCancelBookingRequest" }
      responses:
        "200":
          description: The updated booking detail.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminBookingDetail" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404":
          description: "`booking_not_found`."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
        "409":
          description: "`invalid_state` — the booking is already completed or cancelled."
          content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived HS256 access token from /v1/auth/{login,register,refresh}.
    cookieAuth:
      type: apiKey
      in: cookie
      name: ftr_session
      description: >
        HttpOnly opaque refresh token, Path=/v1/auth, SameSite=Lax. Set by the
        auth endpoints; the browser sends it automatically to /v1/auth/refresh
        and /v1/auth/logout. Not readable or settable by client JS.
  responses:
    BadRequest:
      description: The request was malformed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: The resource does not exist.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Authentication is missing or invalid.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable identifier (e.g. `not_found`).
            message:
              type: string
              description: Human-readable explanation, safe to surface to users.
...
    Currency:
      type: string
      enum: [UZS, USD]
...
    TeacherKind:
      type: string
      enum: [professional, community]
      description: >
        `professional` — certified/qualified teacher, structured lessons.
        `community` — conversation practice, informal tutoring.
    TeacherSort:
      type: string
      enum: [recommended, price_asc, price_desc, rating_desc]
      default: recommended
...
    TeacherProfileWritable:
      type: object
      description: >
        The caller-editable profile fields shared by create and update. Money
        stays in integer minor units. Server-controlled aggregates (`rating`,
        `review_count`, `lessons_completed`, `student_count`,
        `response_time_hours`, `accepting_students`) and `slug` are never
        accepted here.
      properties:
        display_name: { type: string, minLength: 1 }
        headline: { type: string, minLength: 1 }
        kind: { $ref: "#/components/schemas/TeacherKind" }
        country_code: { type: string, minLength: 1, description: ISO 3166-1 alpha-2 }
        country_name: { type: string, minLength: 1 }
        city: { type: string, minLength: 1 }
        timezone:
          type: string
          minLength: 1
          description: IANA name, e.g. "Asia/Tashkent". Must resolve via the tz database.
        price_per_hour_minor: { type: integer, format: int64, minimum: 0 }
        trial_price_minor:
          type: integer
          format: int64
          minimum: 0
          nullable: true
          description: Omit or null for "no trial lesson offered".
        currency:
          allOf: [{ $ref: "#/components/schemas/Currency" }]
          description: Must be `UZS` (the only supported currency). Defaults to `UZS` on create.
        about: { type: string }
        teaching_style: { type: string }
        avatar_url: { type: string }
        intro_video_url: { type: string }
        video_thumbnail_url: { type: string }
        meeting_url:
          type: string
          description: >
            Default video room for this teacher's lessons. Must be an http(s)
            URL, or empty. Never revealed to students until a lesson is confirmed.
        languages:
          type: array
          items: { $ref: "#/components/schemas/LanguageEntry" }
        focus:
          type: array
          items: { type: string, minLength: 1 }
        experience:
          type: array
          items: { $ref: "#/components/schemas/Experience" }
...
    TeacherListResult:
      type: object
      required: [teachers, total, facets]
      properties:
        teachers:
          type: array
          items: { $ref: "#/components/schemas/TeacherSummary" }
        total:
          type: integer
          minimum: 0
          description: Total matches across all pages.
        facets:
          type: object
          required: [languages]
          properties:
            languages:
              type: array
              items: { $ref: "#/components/schemas/LanguageFacet" }
...